Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    double n;
+    float n;
     printf("");
     scanf("%.2f",&n);
     printf("You entered: %.2f/n",n);
